# management-api-server/doc/openapi.json
{
  "openapi" : "3.0.1",
  "info" : {
    "description" : "This is a Restful service for operating Apache Cassandra.  You can find out more about the Management API on [Github](http://github.com/k8ssandra/management-api-for-apache-cassandra)",
    "license" : {
      "name" : "Apache 2.0",
      "url" : "http://www.apache.org/licenses/LICENSE-2.0.html"
    },
    "title" : "Management API for Apache Cassandra™",
    "version" : "0.1"
  },
  "paths" : {
...
    "/api/v0/ops/keyspace" : {
      "get" : {
        "operationId" : "listKeyspaces",
        "parameters" : [ {
          "in" : "query",
          "name" : "keyspaceName",
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "application/json" : {
                "example" : [ "system_schema", "system", "system_auth", "system_distributed", "system_traces" ],
                "schema" : {
                  "type" : "array",
                  "items" : {
                    "type" : "string"
                  }
                }
              }
            },
            "description" : "List of Keyspaces"
          }
        },
        "summary" : "List the keyspaces existing in the cluster"
      }
    },
...
    "/api/v0/ops/keyspace/cleanup" : {
      "post" : {
        "operationId" : "cleanup",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/KeyspaceRequest"
              }
            }
          }
        },
        "responses" : {
          "200" : {
            "content" : {
              "text/plain" : {
                "example" : "d69d1d95-9348-4460-95d2-ae342870fade",
                "schema" : {
                  "type" : "string"
                }
              }
            },
            "description" : "Job ID for keyspace cleanup process"
          }
        },
        "summary" : "Triggers the immediate cleanup of keys no longer belonging to a node. By default, clean all keyspaces. This operation is blocking and will return the executed job after finishing."
      }
    },
...
    "/api/v0/ops/node/decommission" : {
      "post" : {
        "operationId" : "decommission",
        "parameters" : [ {
          "in" : "query",
          "name" : "force",
          "schema" : {
            "type" : "boolean"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "text/plain" : {
                "example" : "OK",
                "schema" : {
                  "type" : "string"
                }
              }
            },
            "description" : "Cassandra node decommissioned successfully"
          }
        },
        "summary" : "Decommission the *node I am connecting to*"
      }
    },
...
    "/api/v1/ops/keyspace/cleanup" : {
      "post" : {
        "operationId" : "cleanup_v1",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/KeyspaceRequest"
              }
            }
          }
        },
        "responses" : {
          "200" : {
            "content" : {
              "text/plain" : {
                "example" : "OK",
                "schema" : {
                  "type" : "string"
                }
              }
            },
            "description" : "Cleanup not needed for tables 'system' or 'system_schema'"
          },
          "202" : {
            "content" : {
              "text/plain" : {
                "example" : "d69d1d95-9348-4460-95d2-ae342870fade",
                "schema" : {
                  "type" : "string"
                }
              }
            },
            "description" : "Job ID for keyspace cleanup process"
          }
        },
        "summary" : "Triggers the immediate cleanup of keys no longer belonging to a node. By default, clean all keyspaces. This operation is asynchronous and returns immediately"
      }
    },
    "/api/v1/ops/node/decommission" : {
      "post" : {
        "operationId" : "decommission_v1",
        "parameters" : [ {
          "in" : "query",
          "name" : "force",
          "schema" : {
            "type" : "boolean"
          }
        } ],
        "responses" : {
          "202" : {
            "content" : {
              "text/plain" : {
                "example" : "34034d36-3c1e-4bdb-8a8f-f92291a64cb3",
                "schema" : {
                  "type" : "string"
                }
              }
            },
            "description" : "Job ID for successfully scheduled Cassandra node decommission request"
          }
        },
        "summary" : "Decommission the *node I am connecting to*. This invocation returns immediately and returns a job id."
      }
    },
    "/api/v1/ops/node/rebuild" : {
      "post" : {
        "operationId" : "rebuild_v1",
        "parameters" : [ {
          "in" : "query",
          "name" : "src_dc",
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "202" : {
            "content" : {
              "text/plain" : {
                "example" : "34034d36-3c1e-4bdb-8a8f-f92291a64cb3",
                "schema" : {
                  "type" : "string"
                }
              }
            },
            "description" : "Job ID for successfully scheduled Cassandra node rebuild request"
          }
        },
        "summary" : "Rebuild data by streaming data from other nodes. This operation returns immediately with a job id."
      }
    },
...
  },
  "components" : {
    "schemas" : {
...
      "KeyspaceRequest" : {
        "type" : "object",
        "properties" : {
          "jobs" : {
            "type" : "integer",
            "format" : "int32"
          },
          "keyspace_name" : {
            "type" : "string"
          },
          "tables" : {
            "type" : "array",
            "items" : {
              "type" : "string"
            }
          }
        },
        "required" : [ "jobs", "keyspace_name", "tables" ]
      },
...
    }
  }
}